requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
crispy-tailwind==0.2.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-environ==0.4.5
django-fontawesome-5==1.0.18
django-js-asset==1.2.2
django-tinymce==3.2.0
django==3.1.5
pillow==8.1.0
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1; python_version >= '3.5'
